requirements.txt
transformers>=4.30.0
torch>=2.0.0
Flask>=2.3.0
Pillow>=10.0.0
gunicorn>=21.0.0
safetensors>=0.3.0
numpy>=1.24.0
python-dotenv>=1.0.0